+++ src/extension.ts
@@ -2,7 +2,14 @@ import * as vscode from 'vscode';
 
 import createComponent from './createComponent';
 
-const handleCreateComponent = async (args: any, styled?: boolean, mobile?: boolean) => {
+interface CreateComponentProps {
+  args: any;
+  named: boolean;
+  styled?: boolean;
+  mobile?: boolean;
+}
+
+const handleCreateComponent = async ({ args, named, styled, mobile }: CreateComponentProps) => {
   const componentName = await vscode.window.showInputBox({
     prompt: `Enter the component name:`,
     ignoreFocusOut: true,
@@ -15,25 +22,31 @@ const handleCreateComponent = async (args: any, styled?: boolean, mobile?: boole
 
   if (args) {
     const path = args.fsPath;
-    createComponent(componentName, { dir: path, styled, mobile });
+    createComponent(componentName, { dir: path, named, styled, mobile });
   } else {
-    createComponent(componentName, { styled, mobile });
+    createComponent(componentName, { named, styled, mobile });
   }
 };
 
 export function activate(context: vscode.ExtensionContext) {
   let disposable = [
     vscode.commands.registerCommand("extension.create-react-component", args => {
-      handleCreateComponent(args);
+      handleCreateComponent({ args, named: false, styled: false, mobile: false });
     }),
     vscode.commands.registerCommand("extension.create-react-styled-component", args => {
-      handleCreateComponent(args, true);
+      handleCreateComponent({ args, named: false, styled: true, mobile: false });
+    }),
+    vscode.commands.registerCommand("extension.create-react-named-component", args => {
+      handleCreateComponent({ args, named: true, styled: false, mobile: false });
     }),
     vscode.commands.registerCommand("extension.create-react-native-component", args => {
-      handleCreateComponent(args, false, true);
+      handleCreateComponent({ args, named: false, styled: false, mobile: true });
     }),
     vscode.commands.registerCommand("extension.create-react-native-styled-component", args => {
-      handleCreateComponent(args, true, true);
+      handleCreateComponent({ args, named: false, styled: true, mobile: true });
+    }),
+    vscode.commands.registerCommand("extension.create-react-native-named-component", args => {
+      handleCreateComponent({ args, named: true, styled: false, mobile: true });
     })
   ];