+++ src/templates/interfaces/CreateComponent.ts
@@ -1,0 +1,7 @@
+export default interface CreateComponent {
+  componentName: string;
+  styleName?: string;
+  useReactImport?: boolean;
+  useReactFC?: boolean;
+  useCSSModule?: boolean;
+}