+++ src/templates/javascript/reactNativeArrowComponent.ts
@@ -1,6 +1,8 @@
-export default (componentName: string) => (
-`import React from 'react';
-import { View, Text, StyleSheet } from 'react-native';
+import CreateComponent from '../interfaces/CreateComponent';
+import creatReactImport from '../shared/functions/create-react-import';
+
+export default ({ componentName, useReactImport }: CreateComponent) => (
+`${ creatReactImport(useReactImport) }import { View, Text, StyleSheet } from 'react-native';
 
 const ${ componentName } = () => {
   return (