+++ src/templates/javascript/reactNativeComponent.ts
@@ -1,10 +1,10 @@
 import CreateComponent from '../interfaces/CreateComponent';
 import creatReactImport from '../shared/functions/create-react-import';
 
-export default ({ componentName, useReactImport, useReactFC }: CreateComponent) => (
+export default ({ componentName, useReactImport }: CreateComponent) => (
 `${ creatReactImport(useReactImport) }import { View, Text, StyleSheet } from 'react-native';
 
-const ${ componentName }${ useReactFC ? ': React.FC' : '' } = () => {
+function ${ componentName }() {
   return (
     <View style={ styles.container }>
       <Text style={ styles.title }>${ componentName }</Text>