+++ src/templates/javascript/styledReactNativeArrowComponent.ts
@@ -1,6 +1,8 @@
-export default (componentName: string) => (
-`import React from 'react';
-import { Text } from 'react-native';
+import CreateComponent from '../interfaces/CreateComponent';
+import creatReactImport from '../shared/functions/create-react-import';
+
+export default ({ componentName, useReactImport }: CreateComponent) => (
+`${ creatReactImport(useReactImport) }import { Text } from 'react-native';
 
 import { Container } from './styles';