+++ src/templates/shared/functions/create-react-import.ts
@@ -1,0 +1,7 @@
+export default function creatReactImport(useReactImport?: boolean) {
+  if (useReactImport) {
+    return `import React from 'react';\n\n`
+  }
+
+  return '';
+}