+++ src/templates/shared/functions/create-styles-import.ts
@@ -1,0 +1,11 @@
+export default function createStylesImport(styleName?: string, useCSSModule?: boolean) {
+  if (styleName === 'styles') {
+    return `import { Container } from './${ styleName }';`
+  } else {
+    if (useCSSModule) {
+      return `import styles from './${ styleName }';`
+    } else {
+      return `import './${ styleName }';`
+    }
+  }
+}