+++ src/templates/typescript/reactComponent.ts
@@ -2,7 +2,7 @@ import CreateComponent from '../interfaces/CreateComponent';
 import creatReactImport from '../shared/functions/create-react-import';
 
 export default ({ componentName, useReactImport }: CreateComponent) => (
-`${ creatReactImport(useReactImport) }const ${ componentName } = () => {
+`${ creatReactImport(useReactImport) }function ${ componentName }() {
   return (
     <h1>${ componentName }</h1>
   );