+++ src/templates/typescript/styledReactComponent.ts
@@ -2,10 +2,10 @@ import CreateComponent from '../interfaces/CreateComponent';
 import creatReactImport from '../shared/functions/create-react-import';
 import createStylesImport from '../shared/functions/create-styles-import';
 
-export default ({ componentName, styleName, useReactImport, useCSSModule }: CreateComponent) => (
+export default ({ componentName, styleName, useReactImport, useReactFC, useCSSModule }: CreateComponent) => (
 `${ creatReactImport(useReactImport) }${ createStylesImport(styleName, useCSSModule) }
 
-const ${ componentName } = () => {
+function ${ componentName }() {
   return (
     ${ styleName === 'styles' ? `<Container>` : `<>` }
       <h1>${ componentName }</h1>