+++ src/templates/typescript/styledReactNativeArrowComponent.ts
@@ -1,10 +1,12 @@
-export default (componentName: string) => (
-`import React from 'react';
-import { Text } from 'react-native';
+import CreateComponent from '../interfaces/CreateComponent';
+import creatReactImport from '../shared/functions/create-react-import';
+
+export default ({ componentName, useReactImport, useReactFC }: CreateComponent) => (
+`${ creatReactImport(useReactImport) }import { Text } from 'react-native';
 
 import { Container } from './styles';
 
-const ${ componentName }: React.FC = () => {
+const ${ componentName }${ useReactFC ? ': React.FC' : '' } = () => {
   return (
     <Container>
       <Text>${ componentName }</Text>