+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to the Create React JS and React Native TSX Component extension will be documented in this file.
 
+## [1.5.0]
+
+- Adds **Orange Icon** 🔶 for React JS Component Actions.
+- Adds **Blue Icon** 🔷 for React Native Component Actions.
+- Adds **Create React JS Named Component** for creation of a React JS Component file with a name of component without folder creation.
+- Adds **Create React Native Named Component** for creation of a React Native Component file with a name of component without folder creation.
+- Adds the setting **useArrowFunctionComponent** that defines if component will be created using an `Arrow Function` or a `Regular Function`.
+- Adds the setting **useReactFC** that defines if component will be typed using `React.FC`. **Important:** Only applicable if option `Use Arrow Function Component` and `Use React Import` is enabled.
+- Adds the setting **useReactImport** that defines if the import of React (`import React from 'react';`) will be used at the beginning of the component. This option should only be used if you are using **React 17+**, that comes with support for the new JSX transform, where we don’t need to import React to our components anymore to use JSX.
+- Adds the setting **useCSSModule** that defines if CSS file name will be sufixed with `.module.css` or `.module.scss`. **Important:** Only applicable if option `Styles Format` is `CSS` or `SCSS`.
+
 ## [1.4.3]
 
 - Fixed the `.css` file extension when uses `.js` files format.